task_9/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

# RAG System Dependencies
azure-search-documents
azure-core
azure-identity
openai>=1.0.0
langchain-openai
langchain
python-dotenv


# Optional: Streamlit Frontend Dependencies
# Install these if you want to run the web frontend:
# streamlit>=1.28.0
# requests>=2.31.0  (already included above)
# pandas>=2.0.0
